:boom: change: introduced ownership semantics for instantiated modex items.
- Ownership represents what module the cached item belongs to.
- Ownership can be used to deduce context for notifications, events,
etc.
- Declared in the pre-compiled header to be used globally.

+++ src/ui/components/UICustom.cpp
@@ -168,7 +168,7 @@ namespace Modex::UICustom
 		return changed;
 	}
 
-	bool FancyDropdown(const char* a_id, const char* a_tooltip, uint8_t& a_currentItem, const std::vector<std::string>& a_items, float a_width)
+	bool FancyDropdown(const char* a_id, const char* a_tooltip, uint32_t& a_currentItem, const std::vector<std::string>& a_items, float a_width)
 	{
 		auto tempIndex = static_cast<int>(a_currentItem);
 		return FancyDropdown(a_id, a_tooltip, tempIndex, a_items, a_width) && (a_currentItem = static_cast<uint8_t>(tempIndex), true);
